separate ACM integration tests
and adapt them for kibana being available by default
