+++ tests/system/config/apm-server.yml.j2
@@ -77,8 +77,10 @@ apm-server:
   ilm.enabled: {{ ilm_enabled }}
   {% endif %}
 
-  {% if kibana_enabled %}
-  kibana.enabled: {{ kibana_enabled }}
+  {% if kibana_enabled is not none or kibana_host is not none %}
+  kibana:
+    {% if kibana_enabled is not none %}enabled: {{ kibana_enabled }}{% endif %}
+    {% if kibana_host is not none %}host: {{ kibana_host }}{% endif %}
   {% endif %}
